extract the headers out of app.js in a new file called Header.js
