+++ App.js
@@ -1,24 +1,14 @@
 import { StatusBar } from 'expo-status-bar';
 import { StyleSheet, Text, View } from 'react-native';
+import Header from "./Components/Header";
 
 export default function App() {
-  const appName = "123 APP";
+  const appName = "Axel's APP";
   return (
-    <View style={styles.container}>
-      <Text>Welcome to {appName}</Text>
-      <StatusBar style="auto" />
-    </View>
+      <Header/>
   );
 }
 
 
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#fff',
-    alignItems: 'center',
-    justifyContent: 'center',
-  },
-});
