+++ Components/Header.js
@@ -1,0 +1,22 @@
+import { StatusBar } from 'expo-status-bar';
+import React from 'react';
+import { StyleSheet, Text, View } from 'react-native';
+
+export default function Header() {
+    const appName = "Axel's APP";
+    return (
+        <View style={styles.container}>
+        <Text>Welcome to {appName}</Text>
+        <StatusBar style="auto" />
+        </View>
+    );
+}
+
+const styles = StyleSheet.create({
+  container: {
+    flex: 1,
+    backgroundColor: '#fff',
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+});
